Rename cachable to cacheable
